luv : os-release : Introduce a numeric BUILD_ID
This patch is in preparation to introduce telemetrics.The telemetrics-client
requires the VERSION_ID to be a number, else records are not transmitted to
the telemetrics server. Currently, VERSION_ID provided by LUV has some special
symbols; hence the records do not reach the server. We have introduced a new
BUILD_ID in the os-release recipe and given it a random numeric value which
will replace the VERSION_ID. Later, the telemetrics server can use this ID to
filter out the records generated by LUV.

In this patch, we have overridden the do_compile function. This has the
original contents of the ..bb file +/- changes required for to add a
numeric BUILD-ID.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>