luv : Introduce telemetrics in LUV
LUV(Linux UEFI Validation) brings together multiple separate upstream
test suites into a cohesive and easy-to-use product and validates UEFI
firmware at critical levels of the software stack. It may be possible
that one of the test suites crashes while running. It may be even
possible that a kernel panic is observed. Under these scenarios, it
would be useful for LUV to call home and submit forensic data to analyze
and address the problem. Of course, this will be an opt-in feature and
users will get clear indication that data is being collected.

We have used the telemetrics-client code developed by the Clear Linux
team and tried to incorporate it in LUV. It has support for crashprobe,
oopsprobe and pstore-probe.

This patch consists of:

1. A yocto recipe which fetches the Clear Linux telemetrics-client repo,
builds it and installs all the necessary service files, daemons and
probes.
2. Add a kernel parameter (luv.telemetrics)which lets the user opt-in to the
telemetrics feature. By default, this feature is disabled.

Currently, this feature is only supported for x86 systems and records
can be sent only fron within intel network.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>