systemd : Enable networking
After shifting to systemD, the LUV image is not being assigned an
IP on boot. This is because it is still using a systemV startup script
to do the same. Since systemD names its interfaces differently, we could
not see any interfaces with a valid IP.
This patch adds the networkd package, introduces a network config file
(50-oe-core.network), which starts dhcp by default for all interfaces
whose names start with en(pci devices which get renamed by udev) or eth
(backward compatible) and a service file (networking.service) which will
bring up the network and make sure an IP is assigned during boot.
It refers:
https://wiki.archlinux.org/index.php/systemd-networkd

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>