linux-yocto-efi-test: enable pstore file system
This patch is in preparation to introduce telemetrics. The telemetrics
client has the pstore probe through which inspects the pstore for oops
messages dropped by dying kernels, creats a telemetry record and sweeps
the pstore clean. Hence, if we ever see a kernel panic when LUV is
running, the last few logs will be store in pstore and when LUV reboots
through kexec, the pstore record sent to the telemetric server can be
investigated for further analysis. By default the pstore file system is
not enabled in the kernel. This patch adds the required kernel configs
as a config fragment.

Also, we have added the crash_kexec_post_notifiers option to make sure
that kdump runs first after the crash followed by kexec and not
vice-versa. Hence, on a crash, the dump is first stored in the pstore
backend and then kexec takes over to reboot the system. This is useful
when there is information in the pstore file system which might be
erased if kexec runs first.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>