luv : increase kernel crash memory
Currently, the crash kernel allocated is 256 MB which is very less to
store the reboot kernel. Hence, we update the crashkernel command line
parameter to 512MB. Also, a contiguous larger memory(512MB) is not
found always in the normal memory, hence we allocate it in the high
memory region. The corresponding kernel configs to allocate memory in
the high memory is added as a separate config fragment.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>